build 0103
The base of the game is finished. Next comes the levels. This means that
all the content for the game is finished. All the game objects,
powerups, characters, sprites, and the general engine/framework of the
game is complete.

After adding the code for two more levels which are complete and
thorough enough and condition for those levels, the (small) game will be
complete.